Thymeleaf

+++ EJERCICIOS/ALEJANDRO_JIMENEZ/demo/src/main/java/com/spring/repositories/ClaseRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.spring.entities.ClaseEntity;
 
-public interface ClaseRepository extends CrudRepository<ClaseEntity, Integer> {
+public interface ClaseRepository extends CrudRepository<ClaseEntity, Long> {
