+++ EJERCICIOS/ALEJANDRO_JIMENEZ/demo/src/main/java/com/spring/repositories/ProfesorRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.spring.entities.ProfesorEntity;
 
-public interface ProfesorRepository extends CrudRepository<ProfesorEntity, Integer> {
+public interface ProfesorRepository extends CrudRepository<ProfesorEntity, Long> {
 	
 @Query("select p from ProfesorEntity p where p.email = :email and p.ciudad = :ciudad order by p.nombre asc,p.ciudad desc")
 List<ProfesorEntity> filtrarEmailCiudad( @Param("email")String email,@Param("ciudad")String ciudad);
